TEST: Line follower
Created a simulation with 3 line followers. The left/right motor speed
is calculated as a linear function of the 3 sensor readings:
Left = 11 + 12*left + 13*middle + 14*right
Right = 21 + 22*left + 23*middle + 24*right

The parameters are passed to each robot from Python as FloatSignals.
This is good groundwork for an optimization algorithm. What's left is to
calculate the performance for each robot at the end of the simulation.